Refactor arrayUtils; add writers imgs
